move search history item and media filters to params package inside usecase because it's not an entity
