--- domain/vod/src/main/java/entity/MediaFilters.kt
@@ -1,7 +1,0 @@
-package entity
-
-data class MediaFilters(
-    val releaseYear: Int? = null,
-    val genres: List<String>? = null,
-    val imdbRating: Float? = null
-)
