--- domain/vod/src/main/java/entity/SearchCategory.kt
@@ -1,7 +1,0 @@
-package entity
-
-enum class SearchCategory {
-    MOVIE,
-    TV_SERIES,
-    ACTOR
-}
