--- domain/vod/src/main/java/entity/SearchHistoryItem.kt
@@ -1,11 +1,0 @@
-package entity
-
-import kotlinx.datetime.LocalDateTime
-
-data class SearchHistoryItem(
-    val query: String,
-    val category: SearchCategory,
-    val filters: MediaFilters? = null,
-    val language: Language,
-    val timestamp: LocalDateTime
-)
