+++ domain/vod/src/main/java/usecase/SearchMoviesUseCase.kt
@@ -4,10 +4,10 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 import entity.Language
-import entity.MediaFilters
+import usecase.params.MediaFilters
 import entity.Movie
-import entity.SearchCategory
-import entity.SearchHistoryItem
+import usecase.params.SearchCategory
+import usecase.params.SearchHistoryItem
 import repository.SearchHistoryRepository
 import repository.SearchRepository
