+++ domain/vod/src/main/java/usecase/params/SearchCategory.kt
@@ -1,4 +1,4 @@
-package entity
+package usecase.params
 
 enum class SearchCategory {
     MOVIE,
